Feature: Observer Design Pattern
focuses on one to many relationship
This design pattern brings with it a variety of advantages:

The pattern is an excellent example of applying the principle of making the coupling between objects as weak as possible. So, increases the so-called loosely.
It allows data to be sent to other objects effectively without any change in the Subject or Observer classes.
Subjects and Observers are independent, and a change in one of these components does not require a change in the other component.
Observers can be added/removed at any time